twitterv3.rake is working fine

+++ app/models/corpus_anew.rb
@@ -1,5 +1,5 @@
 class CorpusAnew
   include Mongoid::Document
   field :sword, type: String
-  field :valmnall, type: Double
+  field :valmnall, type: BigDecimal
 end
